modules/system-target: Update with modern NixOS/Nixpkgs semantics
It's been a while since I authored this module, and things have changed.

This fixes an issue where an assertion triggers due to setting the
legacy `crossSystem` option. And the assertion (needlessly) checks the
definition's priority to trigger.

So even forcing the original default value will fail!

In turn, this means that configurations attempting to handle cross for
the Mobile configuration cannot compose due to the NixOS assertion.