gui: Don't override libraries for stage-2
Using `replaceDependencies` here saves us a costly rebuild, even though
it's not very "pure" to do so.

I believe it is the better option.

This fixes an issue where `libxkbcommon` in stage-2 didn't know about
the usual xkeyboard config files.

As a reminder, in stage-1 xkeyboard config files are handled
differently.