modules/initrd-kernel: Fix breakage from NixOS splitting kernel outputs
This handles the breakage in a forward-and-backward compatible manner,
by looking at if `modules` is available. It will be assumed that if it
is, that it's what should be used.

I have tested by using the changes in https://github.com/NixOS/nixpkgs/pull/436304
to confirm that an empty closure was not produced.